# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Check if API is working",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/wordcount": {
            "post": {
                "description": "Count freuqent words in text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "count"
                ],
                "summary": "Count freuqent words in text",
                "parameters": [
                    {
                        "description": "wordcountrequest",
                        "name": "wordcountrequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.WordCountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.WordCountResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/wordcount/file": {
            "post": {
                "description": "Count freuqent words in text file (.txt file)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "count"
                ],
                "summary": "Count freuqent words in text file (.txt file)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "text file",
                        "name": "myFile",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.WordCountResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.WordCountRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "controllers.WordCountResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.WordCount"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.WordCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "word": {
                    "type": "string"
                }
            }
        }
    }
}